requirements.txt
# PaperTrail — Requirements
# pip install -r requirements.txt

# ── UI ─────────────────────────────────────────────────────────────────────────
streamlit>=1.35.0

# ── LLM ────────────────────────────────────────────────────────────────────────
anthropic>=0.49.0

# ── Embeddings — Google Gemini ─────────────────────────────────────────────────
google-generativeai>=0.8.0

# ── Reranker — Cohere ──────────────────────────────────────────────────────────
cohere>=5.0.0

# ── Vector store ───────────────────────────────────────────────────────────────
chromadb>=0.5.0

# ── Keyword search ─────────────────────────────────────────────────────────────
rank-bm25>=0.2.2

# ── PDF parsing ────────────────────────────────────────────────────────────────
pymupdf>=1.23.0

# ── Utilities ──────────────────────────────────────────────────────────────────
python-dotenv>=1.0.0

# ── Optional OCR (for scanned PDFs) ───────────────────────────────────────────
# pytesseract>=0.3.10
# Pillow>=10.0.0
# Also run: sudo apt install tesseract-ocr
